Criado proteção pata views and forms
close #26
