+++ templates/login.html
@@ -3,26 +3,27 @@
 {% block title %}Log in to Django Boards{% endblock %}
 
 {% block content %}
-  <div class="row justify-content-center">
-    <div class="col-lg-4 col-md-6 col-sm-8">
-      <div class="card">
-        <div class="card-body">
-          <h3 class="card-title">Log in</h3>
-          <form method="post" novalidate>
-            {% csrf_token %}
-            {% include 'includes/form.html' %}
-            <button type="submit" class="btn btn-primary btn-block">Log in</button>
-          </form>
+    <div class="row justify-content-center">
+        <div class="col-lg-4 col-md-6 col-sm-8">
+            <div class="card">
+                <div class="card-body">
+                    <h3 class="card-title">Log in</h3>
+                    <form method="post" novalidate>
+                        {% csrf_token %}
+                        <input type="hidden" name="next" value="{{ next }}">
+                        {% include 'includes/form.html' %}
+                        <button type="submit" class="btn btn-primary btn-block">Log in</button>
+                    </form>
+                </div>
+                <div class="card-footer text-muted text-center">
+                    New to Django Boards? <a href="{% url 'signup' %}">Sign up</a>
+                </div>
+            </div>
+            <div class="text-center py-2">
+                <small>
+                    <a href="#" class="text-muted">Forgot your password?</a>
+                </small>
+            </div>
         </div>
-        <div class="card-footer text-muted text-center">
-          New to Django Boards? <a href="{% url 'signup' %}">Sign up</a>
-        </div>
-      </div>
-      <div class="text-center py-2">
-        <small>
-          <a href="#" class="text-muted">Forgot your password?</a>
-        </small>
-      </div>
     </div>
-  </div>
 {% endblock %}
